+++ app/src/main/java/com/ktx/dormitory/student/payment/data/dto/request/OnlinePaymentRequestDto.kt
@@ -1,6 +1,7 @@
 package com.ktx.dormitory.student.payment.data.dto.request
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 /**
  * Request DTO cho việc tạo thanh toán online (Smart QR).
@@ -8,7 +9,7 @@ import com.google.gson.annotations.SerializedName
  */
 data class OnlinePaymentRequestDto(
     @SerializedName("billId") val billId: String,
-    @SerializedName("amount") val amount: Double,
+    @SerializedName("amount") val amount: BigDecimal,
     @SerializedName("paymentMethod") val paymentMethod: String = "BANK_TRANSFER",
     @SerializedName("transactionCode") val transactionCode: String? = null
 )
