+++ app/src/main/java/com/ktx/dormitory/student/payment/data/remote/PaymentRemoteDataSource.kt
@@ -5,11 +5,12 @@ import com.ktx.dormitory.core.common.PageResponse
 import com.ktx.dormitory.student.payment.data.dto.response.BillDto
 import com.ktx.dormitory.student.payment.data.dto.response.PaymentResponseDto
 import com.ktx.dormitory.student.payment.data.dto.response.PaymentInstructionDto
+import java.math.BigDecimal
 
 interface PaymentRemoteDataSource {
     suspend fun getBillByApplication(applicationId: String): BaseResponse<BillDto>
     suspend fun getInvoices(): BaseResponse<List<BillDto>>
-    suspend fun createSmartQR(billId: String, amount: Double, paymentMethod: String, transactionCode: String?): BaseResponse<PaymentResponseDto>
+    suspend fun createSmartQR(billId: String, amount: BigDecimal, paymentMethod: String, transactionCode: String?): BaseResponse<PaymentResponseDto>
     suspend fun getPaymentHistoryPaged(page: Int, size: Int): BaseResponse<PageResponse<BillDto>>
     suspend fun getPaymentInstructions(): BaseResponse<PaymentInstructionDto>
 }
